Advanced Efficient Audit Service in Cloud Storage
Our primary goal was to develop a solution that will enable the client
to check the integrity of its remote data at any time. As we carried on
with the development of our initial idea, we realized that the solution
must not be computationally too demanding. The efficiency gains were
part of the cloud computing benefits to persuade us on outsourcing data
and services to a third party, at the first place. Thus, we may
conclude that the solution must not, under any circumstance, download
the outsourced file or read it entirely.